A different way of upgrading built-in models
It's possible to upgrade a built-in model without creating a new model
that treats the old model as its base. Basically, you don't create a new
model, you create a new boot-script. The boot-script should be loaded
ASAP, so it's probably best to prefix it with a number. Then in your
boot-script you grab a reference to the model you want to alter, alter
it, and then go have a nap.

The solution was discovered (here)[extending-models].

[extending-models]: https://gist.github.com/pulkitsinghal/43d4bae2467c686ec55b